feat: Adds CRUD for policies
Fixes #31. Note that the update functionality doesn't include action and command. This is because postgres expects a DROP and recreate to alter these settings. I don't want to do that within the PATCH command because it will return a new ID. End users should follow the convention - delete and recreate