docs(openapi): add initial OpenAPI spec
This adds the initial OpenAPI spec for pg-api.

Most of the skeleton is in place, but it's still missing some things,
e.g.:

- Non-boiletplate examples
- Failing HTTP responses
- Descriptions
- Test against an OpenAPI validator
- Verify required/nullable fields against actual properties on Postgres

I don't think we can get this right the first time, so let's go with
this and Kaizen the rest.

Also, this spec slightly differs from the current
implementation (camelCase vs. snake_case, type inconsistencies). These
will be addressed in the future.