feat(lib): always use pg-pool, support pool.end()
* Always use `Pool` from node-postgres.

Originally the library allows using the `Client` class from
node-postgres (by setting `pooled: false`) which is allegedly for when
we only need one DB connection that we manage manually, however it turns
out that using it directly is not the "happy path", and it's simpler to
just use a `Pool` with max size 1 (`pool.end()` ends connections
gracefully, among other things--`client.end()` ends ongoing queries
abruptly).

* Support `pool.end()` from the client.

Since we're dynamically creating pools in the server, we want the
library to be able to end unneeded pools. Note that in the typical use
case of one DB connection this isn't necessary--pools are closed when
the server stops.