Skip over dropped attributes when enumerating types
`pg_attribute` may contain records for attributes that were dropped but Postgres kept them around and instead of deleting them, renamed them to `........pg.dropped.#........` and set their `attisdropped` to `true`.

This ends up generating these dropped attributes as `unknown` fields in the TypeScript types.

In this commit I updated the `types` query to not return these.

Maybe it would be preferrable to instead keep these but skip them only when generating the TypeScript types?