Allow strings for `numeric` type inserts and updates
When generating TypeScript types, the current code "correctly" reports
numeric types as `number`, as Postgres and Postgrest return such columns
as JSON numbers by default without a query `::text` cast.

However, generated insert and update types are limited to `number`,
whereas the underlying APIs will accept a string and allow preserving
precision that would be lost in conversion to JS `number` this way.

Generated types for inserting and updating numeric fields now include
`| string` as a possibility to allow for this.